change preview modal image to original image
