+++ src/layout/FileUpload/FileUploadTable/AttachmentThumbnail.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
 
 import { isAttachmentUploaded } from 'src/features/attachments';
+import { useInstanceDataElements, useLaxInstanceId } from 'src/features/instance/InstanceContext';
+import { useCurrentLanguage } from 'src/features/language/LanguageProvider';
 import classes from 'src/layout/FileUpload/FileUploadTable/AttachmentThumbnail.module.css';
-import { useThumbnailLink } from 'src/layout/FileUpload/FileUploadTable/useThumbnailLink';
+import { getDataElementUrl } from 'src/utils/urls/appUrlHelper';
+import { makeUrlRelativeIfSameDomain } from 'src/utils/urls/urlHelper';
 import type { IAttachment } from 'src/features/attachments';
 
 interface IAttachmentThumbnailProps {
@@ -12,12 +15,29 @@ interface IAttachmentThumbnailProps {
 }
 
 export const AttachmentThumbnail = ({ attachment, mobileView, onThumbnailClick }: IAttachmentThumbnailProps) => {
-  const thumbnailUrl = useThumbnailLink(attachment);
+  const dataElements = useInstanceDataElements(undefined);
+  const instanceId = useLaxInstanceId();
+  const language = useCurrentLanguage();
 
   // Only uploaded attachments can have thumbnails
-  if (!isAttachmentUploaded(attachment) || !thumbnailUrl) {
-    return null;
+  if (!instanceId || !isAttachmentUploaded(attachment)) {
+    return '';
   }
+
+  const thumbnailLink = attachment.data.metadata?.find((meta) => meta.key === 'thumbnailLink')?.value;
+  if (!thumbnailLink) {
+    return '';
+  }
+
+  const thumbnailDataElement = dataElements.find(
+    (el) =>
+      el.dataType === 'thumbnail' &&
+      el.metadata?.some((meta) => meta.key === 'attachmentLink' && meta.value === thumbnailLink),
+  );
+
+  const url = thumbnailDataElement
+    ? makeUrlRelativeIfSameDomain(getDataElementUrl(instanceId, thumbnailDataElement.id, language))
+    : undefined;
 
   return (
     <div
@@ -30,7 +50,7 @@ export const AttachmentThumbnail = ({ attachment, mobileView, onThumbnailClick }
       onKeyDown={onThumbnailClick ? (e) => e.key === 'Enter' && onThumbnailClick() : undefined}
     >
       <img
-        src={thumbnailUrl}
+        src={url}
         alt={`Thumbnail for ${attachment.data.filename}`}
         className={mobileView ? classes.thumbnailMobile : classes.thumbnail}
       />
